Refine API for immutable factory functions
